training/requirements.txt
kagglehub
uvicorn
tensorflow
azure.storage.blob
python-multipart